Fixed Broken tests and separated out the components into their own tests. Also fixed a small logging issue.
