+++ src/main/java/com/synopsys/integration/alert/workflow/startup/component/ConfigurationLogger.java
@@ -1,3 +1,25 @@
+/**
+ * blackduck-alert
+ *
+ * Copyright (c) 2019 Synopsys, Inc.
+ *
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements. See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership. The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
 package com.synopsys.integration.alert.workflow.startup.component;
 
 import java.util.Optional;
@@ -15,12 +37,11 @@ import com.synopsys.integration.alert.common.persistence.accessor.FieldAccessor;
 import com.synopsys.integration.alert.common.persistence.model.ConfigurationModel;
 import com.synopsys.integration.alert.provider.blackduck.BlackDuckProperties;
 import com.synopsys.integration.alert.provider.blackduck.descriptor.BlackDuckDescriptor;
-import com.synopsys.integration.alert.workflow.startup.StartupManager;
 
 @Component
 @Order(3)
 public class ConfigurationLogger extends StartupComponent {
-    private final Logger logger = LoggerFactory.getLogger(StartupManager.class);
+    private final Logger logger = LoggerFactory.getLogger(ConfigurationLogger.class);
 
     private final ProxyManager proxyManager;
     private final BlackDuckProperties blackDuckProperties;
