+++ src/test/java/com/synopsys/integration/alert/workflow/startup/StartupTaskTest.java
@@ -1,0 +1,4 @@
+package com.synopsys.integration.alert.workflow.startup;
+
+public class StartupTaskTest {
+}
